Replace original dtfconcat tool with new version
